+++ lib/src/movements.dart
@@ -8,5 +8,10 @@ class Movements {
   static final north = new Point(0, -1);
   static final south = new Point(0, 1);
 
+  static final northwest = new Point(-1, -1);
+  static final northeast = new Point(1, -1);
+  static final southwest = new Point(-1, 1);
+  static final southeast = new Point(1, 1);
+
   static final none = new Point(0, 0);
 }